Refactor indicator registration
- Add QuoteTick and TradeTick updating
- Move SpreadAnalyzer to indicators
