+++ examples/strategies/ema_cross_filtered.py
@@ -21,6 +21,7 @@ import pytz
 from nautilus_trader.common.timer import TimeEvent
 from nautilus_trader.indicators.atr import AverageTrueRange
 from nautilus_trader.indicators.average.ema import ExponentialMovingAverage
+from nautilus_trader.indicators.spread_analyzer import SpreadAnalyzer
 from nautilus_trader.model.bar import Bar
 from nautilus_trader.model.bar import BarSpecification
 from nautilus_trader.model.bar import BarType
@@ -30,7 +31,6 @@ from nautilus_trader.model.enums import TimeInForce
 from nautilus_trader.model.identifiers import Symbol
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.tick import QuoteTick
-from nautilus_trader.trading.analyzers import SpreadAnalyzer
 from nautilus_trader.trading.filters import EconomicNewsEventFilter
 from nautilus_trader.trading.filters import ForexSession
 from nautilus_trader.trading.filters import ForexSessionFilter
@@ -132,18 +132,10 @@ class EMACrossFiltered(TradingStrategy):
         self.quote_currency = instrument.quote_currency
 
         # Register the indicators for updating
-        self.register_indicator(
-            data_source=self.bar_type,
-            indicator=self.fast_ema,
-            update_method=self.fast_ema.update)
-        self.register_indicator(
-            data_source=self.bar_type,
-            indicator=self.slow_ema,
-            update_method=self.slow_ema.update)
-        self.register_indicator(
-            data_source=self.bar_type,
-            indicator=self.atr,
-            update_method=self.atr.update)
+        self.register_indicator_for_quote_ticks(self.symbol, self.spread_analyzer)
+        self.register_indicator_for_bars(self.bar_type, self.fast_ema)
+        self.register_indicator_for_bars(self.bar_type, self.slow_ema)
+        self.register_indicator_for_bars(self.bar_type, self.atr)
 
         # Set trading sessions
         self._update_session_times()
@@ -167,7 +159,6 @@ class EMACrossFiltered(TradingStrategy):
         :param tick: The quote tick received.
         """
         # self.log.info(f"Received Tick({tick})")  # For debugging
-        self.spread_analyzer.update(tick)
 
     def on_bar(self, bar_type: BarType, bar: Bar):
         """
